Homework4 mistakes.

+++ src/main/java/ru/x5/hw04/task1/Shape.java
@@ -2,6 +2,6 @@ package ru.x5.hw04.task1;
 
 public interface Shape {
 
-    int square();
+    double square();
 
 }
